+++ tsup.config.ts
@@ -12,6 +12,7 @@ export default defineConfig({
     'precompile/index': 'src/precompile/index.ts',
     'signers/types': 'src/signers/types.ts',
     'signers/node/index': 'src/signers/node/index.ts',
+    'signers/browser/index': 'src/signers/browser/index.ts',
   },
   format: ['esm', 'cjs'],
   dts: true,